Replace ConfigDirectoryRequirement with DirectoryRequirement
refs #2582
